refactor: extract vote form into their own component
