+++ apps/ui/src/components/ProposalVoteBasic.vue
@@ -1,0 +1,48 @@
+<script setup lang="ts">
+import { Choice } from '@/types';
+
+withDefaults(
+  defineProps<{
+    sendingType: Choice | number | number[] | null;
+    size: number;
+  }>(),
+  { size: 48 }
+);
+
+const emit = defineEmits<{
+  (e: 'handleVoteClick', value: Choice);
+}>();
+</script>
+
+<template>
+  <UiTooltip title="For">
+    <UiButton
+      class="!text-skin-success !border-skin-success !px-0"
+      :class="{ '!w-[48px] !h-[48px]': size === 48, '!w-[40px] !h-[40px]': size === 40 }"
+      :loading="sendingType === 'for'"
+      @click="emit('handleVoteClick', 'for')"
+    >
+      <IH-check class="inline-block" />
+    </UiButton>
+  </UiTooltip>
+  <UiTooltip title="Against">
+    <UiButton
+      class="!text-skin-danger !border-skin-danger !px-0"
+      :class="{ '!w-[48px] !h-[48px]': size === 48, '!w-[40px] !h-[40px]': size === 40 }"
+      :loading="sendingType === 'against'"
+      @click="emit('handleVoteClick', 'against')"
+    >
+      <IH-x class="inline-block" />
+    </UiButton>
+  </UiTooltip>
+  <UiTooltip title="Abstain">
+    <UiButton
+      class="!text-gray-500 !border-gray-500 !px-0"
+      :class="{ '!w-[48px] !h-[48px]': size === 48, '!w-[40px] !h-[40px]': size === 40 }"
+      :loading="sendingType === 'abstain'"
+      @click="emit('handleVoteClick', 'abstain')"
+    >
+      <IH-minus-sm class="inline-block" />
+    </UiButton>
+  </UiTooltip>
+</template>
